+++ src/submissions/__main__.py
@@ -1,7 +1,6 @@
-import sys
-import os
-# environment variable must be set to enable qtwebengine in network path
+import sys, os
 from tools import ctx, setup_logger, check_if_app
+# environment variable must be set to enable qtwebengine in network path
 if check_if_app():
     os.environ['QTWEBENGINE_DISABLE_SANDBOX'] = "1"
 # setup custom logger