+++ CHANGELOG.md
@@ -1,7 +1,9 @@
+## 202312.02
+
 ## 202312.01
 
 - Control samples info now available in plate map.
-- Backups will now create an regenerated xlsx file.
+- Backups will now create a regenerated xlsx file.
 - Report generator now does sums automatically.
 
 ## 202311.04